refactor/#27 status 리포지토리 파스칼로 명칭 변경
