+++ domain/repositories/IStatusRepository.ts
