+++ infrastructure/repositories/index.ts
@@ -1,2 +1,2 @@
 export * from "@/infrastructure/repositories/PriUserRepository";
-export * from "@/infrastructure/repositories/pr-status-repository";
+export * from "@/infrastructure/repositories/PriStatusRepository";
